requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
async-timeout==4.0.3
awscli==1.27.79
azure-common==1.1.28
azure-core==1.31.0
azure-identity==1.16.1
azure-mgmt-authorization==4.0.0
azure-mgmt-core==1.4.0
azure-mgmt-loganalytics==12.0.0
azure-mgmt-resource==23.1.0
azure-monitor-query==1.3.0
backoff==2.2.1
bcrypt==4.2.0
billiard==3.6.4.0
boto3==1.26.79
botocore==1.29.79
cachetools==5.5.0
celery==5.2.7
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
clickhouse-connect==0.6.23
clickhouse-driver==0.2.6
clickhouse-pool==0.5.3
colorama==0.4.4
cryptography==43.0.1
datadog-api-client==2.15.0
defusedxml==0.7.1
distro==1.9.0
dj-database-url==1.2.0
dj-email-url==1.0.6
dj-rest-auth==2.2.7
Django==4.1.4
django-allauth==0.52.0
django-appconf==1.0.5
django-cache-url==3.4.4
django-celery-beat==2.4.0
django-celery-email==3.0.0
django-clickhouse-backend==1.0.3
django-cors-headers==3.13.0
django-extensions==3.2.1
django-filter==22.1
django-prometheus==2.2.0
django-redis==5.2.0
django-rest-auth==0.9.5
django-ses==3.3.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docutils==0.16
durationpy==0.7
elastic-transport==8.15.0
elasticsearch==8.15.1
environs==9.5.0
exceptiongroup==1.2.2
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gql==3.5.0
graphql-core==3.2.4
grpcio==1.62.0
grpcio-status==1.62.0
grpcio-tools==1.51.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.10
isodate==0.6.1
jmespath==1.0.1
kombu==5.4.2
kubernetes==26.1.0
lz4==4.3.3
marshmallow==3.22.0
msal==1.31.0
msal-extensions==1.2.0
msrest==0.7.1
multidict==6.1.0
mypy-protobuf==3.4.0
oauthlib==3.2.2
openai==1.35.15
opsgenie-sdk==2.1.5
packaging==24.1
paramiko==3.4.1
pdpyras==5.2.0
portalocker==2.10.1
prometheus_client==0.21.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.21.12
psycopg2-binary==2.9.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyNaCl==1.5.0
pyparsing==3.1.4
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
PyYAML==5.4.1
redis==5.0.8
requests==2.28.2
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.7.2
s3transfer==0.6.2
six==1.16.0
slack_sdk==3.26.2
sniffio==1.3.1
SQLAlchemy==1.4.54
sqlparse==0.5.1
tenacity==9.0.0
tqdm==4.66.5
types-protobuf==5.28.0.20240924
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
yarl==1.12.1
zstandard==0.23.0
